requirements.txt
textual>=0.40.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
rich>=13.6.0
markdown>=3.4.0
